+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Changed
 - `rofimoji` uses [Poetry](https://python-poetry.org/) for package management. With this comes the requirement for Python 3.8.
 
+## Added
+- You can now select your desired actions through a separate menu.
+
 # [5.6.0]
 ## Changed
 - Remove the unexpected and distracting empty checkboxes from rofi. (#127, #125)